adds call for depositing amount

+++ src/main/java/net/worth/banking_app/entity/Account.java
@@ -4,7 +4,6 @@ import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.NonNull;
 
 @Data
 @NoArgsConstructor
@@ -16,6 +15,6 @@ public class Account {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
     private String accountHolderName;
-    private Long balance;
+    private double balance;
 }
